Refactor, move logic from LatestRelease to VersionComparator
