+++ rewrite-core/src/main/java/org/openrewrite/semver/HyphenRange.java
@@ -24,7 +24,7 @@ import java.util.regex.Pattern;
 /**
  * <a href="https://github.com/npm/node-semver#hyphen-ranges-xyz---abc">Hyphen ranges</a>.
  */
-public class HyphenRange extends LatestRelease {
+public class HyphenRange extends VersionComparator {
     private static final Pattern HYPHEN_RANGE_PATTERN = Pattern.compile("(\\d+(\\.\\d+)?(\\.\\d+)?(\\.\\d+)?)\\s*-\\s*(\\d+(\\.\\d+)?(\\.\\d+)?(\\.\\d+)?)");
 
     private final String upper;
